requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.8
greenlet==1.1.3
idna==3.3
iniconfig==1.1.1
packaging==21.3
playwright==1.27.1
pluggy==1.0.0
py==1.11.0
pyee==8.1.0
pyparsing==3.0.6
pytest==7.2.0
pytest-base-url==1.4.2
pytest-playwright==0.2.2
python-dotenv==0.21.0
pytest-dotenv==0.5.2
python-slugify==5.0.2
six==1.16.0
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.7
websockets==10.1
pandas==2.2.2
pytest-rerunfailures==14.0